modularizing page
