+++ components/buttons/fitlerBtns.js
@@ -1,0 +1,15 @@
+import renderToDOM from '../../utils/utility';
+
+const filterBtnRow = () => {
+  const domString = `<div class='btn-group' role='group' aria-label='Basic example'>
+      <button type='button' id='filter--hufflepuff' class='btn btn-warning btn-sm'>Hufflepuff</button>
+      <button type='button' class='btn btn-primary btn-sm' id='filter--ravenclaw'>Ravenclaw</button>
+      <button type='button' class='btn btn-success btn-sm' id='filter--slytherin'>Slytherin</button>
+      <button type='button' class='btn btn-danger btn-sm' id='filter--gryffindor'>Gryffindor</button>
+      <button type='button' class='btn btn-secondary btn-sm' id='filter--all'>All</button>
+    </div>`;
+
+  renderToDOM('#filter-container', domString);
+};
+
+export default filterBtnRow;
