+++ components/shared/createId.js
@@ -1,0 +1,11 @@
+// ********** LOGIC  ********** //
+// Create a new ID for the students
+const createId = (array) => {
+  if (array.length) {
+    const idArray = array.map((el) => el.id);
+    return Math.max(...idArray) + 1;
+  }
+  return 0;
+};
+
+export default createId;
