+++ events/formEvents.js
@@ -1,0 +1,49 @@
+import students from '../utils/data/students';
+import voldysArmy from '../utils/data/voldyArmy';
+import studentsOnDom from '../utils/studentsOnDom';
+import studentAreas from '../components/shared/studentAreas';
+import form from '../components/forms/form';
+import filterBtnRow from '../components/buttons/fitlerBtns';
+
+const events = () => {
+  // get form on the DOM on button click
+  document.querySelector('#start-sorting').addEventListener('click', () => {
+  // put html elements on the DOM on click
+    form(); // form
+    filterBtnRow(); // filter buttons
+    studentAreas(); // students and voldy's army divs
+  });
+  // target expel buttons to move to voldys army
+  document
+    .querySelector('#student-container')
+    .addEventListener('click', (e) => {
+      if (e.target.id.includes('expel')) {
+        const [, id] = e.target.id.split('--');
+        const index = students.findIndex(
+          (student) => student.id === Number(id)
+        );
+
+        // move from one array to another
+        voldysArmy.push(...students.splice(index, 1));
+        // get both sets of students on the DOM
+        studentsOnDom('#students', students);
+        studentsOnDom('#voldy', voldysArmy);
+      }
+    });
+
+  // target filter buttons on Dom
+  document.querySelector('#filter-container').addEventListener('click', (e) => {
+    if (e.target.id.includes('filter')) {
+      const [, house] = e.target.id.split('--');
+
+      if (house === 'all') {
+        studentsOnDom('#students', students);
+      } else if (house) {
+        const filter = students.filter((student) => student.house === house);
+        studentsOnDom('#students', filter, house);
+      }
+    }
+  });
+};
+
+export default events;
