+++ utils/data/students.js
@@ -1,0 +1,2 @@
+const students = [];
+export default students;
