+++ utils/data/voldyArmy.js
@@ -1,0 +1,2 @@
+const voldysArmy = []; // starts as an empty array
+export default voldysArmy;
