+++ utils/utility.js
@@ -1,0 +1,5 @@
+const renderToDOM = (divId, content) => {
+  const selectedDiv = document.querySelector(divId);
+  selectedDiv.innerHTML = content;
+};
+export default renderToDOM;
